llm_core_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
langchain==0.2.7
langchain-google-genai==1.0.10
langchain-community
chromadb>=0.5.0
pysqlite3-binary
requests